docs(Breadcrump): update Breadcrump docs
