+++ src/BreadcrumbItem.js
@@ -4,10 +4,14 @@ import classNames from 'classnames';
 import { mapToCssModules, tagPropType } from './utils';
 
 const propTypes = {
-  tag: tagPropType,
+  /** Adds a visual "active" state to a Breadcrumb Item */
   active: PropTypes.bool,
+  /** Add custom class to the element */
   className: PropTypes.string,
+  /** Change existing className with a new className */
   cssModule: PropTypes.object,
+  /** Set a custom element for this component */
+  tag: tagPropType,
 };
 
 const defaultProps = {
