+++ stories/Breadcrumb.stories.js
@@ -15,6 +15,6 @@ Indicate the current page’s location within a navigational hierarchy that auto
   }
 };
 
-export { default as Breadcrumb } from './examples/Breadcrumb';
-export { default as NoList } from './examples/BreadcrumbNoList';
-export { default as Props } from './examples/BreadcrumbProps';
+export { default as Breadcrumb } from './examples/Breadcrump/Breadcrumb';
+export { default as NoList } from './examples/Breadcrump/BreadcrumbNoList';
+export { default as Props } from './examples/Breadcrump/BreadcrumbProps';
