+++ stories/examples/Breadcrump/BreadcrumbProps.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Breadcrumb, BreadcrumbItem } from 'reactstrap';
-import Props from './Props';
+import Props from '../Props';
 
 const Example = () => (
   <Props components={[Breadcrumb, BreadcrumbItem]} />
